[change] - adding address to user

+++ src/main/java/pl/wolski/bank/services/UserService.java
@@ -4,6 +4,7 @@ package pl.wolski.bank.services;
 import pl.wolski.bank.models.User;
 import org.springframework.security.core.userdetails.UserDetailsService;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface UserService extends UserDetailsService {
@@ -12,5 +13,5 @@ public interface UserService extends UserDetailsService {
 
     boolean isUniqueLogin(String login);
     boolean isUniqueEmail(String email);
-    boolean isUniquePersonalIdentificationNumber(BigInteger personalIdentificationNumber);
+    boolean isUniquePersonalIdentificationNumber(BigDecimal personalIdentificationNumber);
 }
