+++ src/main/java/pl/wolski/bank/services/UserServiceImpl.java
@@ -14,6 +14,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.HashSet;
@@ -83,7 +84,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean isUniquePersonalIdentificationNumber(BigInteger personalIdentificationNumber) {
+    public boolean isUniquePersonalIdentificationNumber(BigDecimal personalIdentificationNumber) {
         return userRepository.findByPersonalIdentificationNumber(personalIdentificationNumber) == null;
     }
 }
